Cleanup app and tests
A lot of the extra stuff we had for logging exceptions and
such aren't required any longer as the WASI bits are working
as they ought to, so we can gut all of that.

Also, fixed vcpu_time which does return an integer number
of milliseconds in the viceroy impl, so assert on that.