testing: export pytest viceroy helpers as part of sdk
We are using these helpers internally and it probably makes sense
to have some sort of easier setup for end users.  This is the first
stab at making that available.  Along with those changes, I also
incorporated a few chanes to reduce the code for viceroy setup
using asyncio (still a bit verbose but is working) and a few
other small tweaks.

The plugin to get Viceroy output on failure is a hack and we may
scrap or have to revisit other approaches (e.g. just immediately
outputting to host stdout/stderr) but it is functional and may
be a reasonable approach (if ugly in implementation).