Flask support!
Here move away from app.py at the top-level in favor of an examples
directory.  We clone the functionality of the bottle example and
port it to flask.  Things don't work out of the box, and we define
a set of hack shims for now that manage to make it so at least
a very basic service can be written in flask!