WIP: backend api experimentation
Here we introduc two new examples; in the first we do some basic work
direclty against the WIT bindings to feel out its edges and understand
its behavior.  In the second example, we use a newly introduced
fastly_compute.requests module that, at this point, provides a
synchronous interface for making backend requests that aims to be
compatible with existing requests package usages (so users may
be able to import our requests module as requets).