Switch to `wasm32-unknown-unknown` target, which will let this link under Viceroy.
I skipped ahead using wit-bindgen to generate all the stubs to see what would happen, and I ran into a link error. It turns out that the `wasm32-wasip2` target adds a lot of extra imports, 0.2.3 WASI ones, which Viceroy doesn't provide and we obviously cannot provide to ourselves. We'll just have to wrap our output in a component ourselves with wasm-tools after building. We'll work that into the build process eventually.