Regularize wac based on an examination of the `compute` package and the imports of `wit-bottle.wasm`.
In other words, this "should work".

Specifically, revise the wac to not import anything that is provided by the Compute `service-imports` world. Those items should be provided directly by the host (Viceroy or similar), as then it will actually be functional.

Add in `poll`, as Python imports it but `service-imports` does not (so it's not provided to services by the host).