Introduce snapshot testing and convert backend tests
I'm familiar with snapshot testing from the rust world
using the insta library.  Here we introduce syrupy
to do snapshot testing to assert that we get back
the responses we want.

For now, this is only applied to the backend tests
which needed some updates as it is.  The general flow
is to add tests and then to run `make test-update-snapshots`
(or run the syrupy command directly) in order to write
out new snapshots (which should be reviewed).