Add pyrefly type checker & lsp support (and fix defects)
Pyrefly is fast and the defaults seem to be a bit less
aggressive than basedpyright.  Ty is another option but
it doesn't seem to be as mature, potentially.
This adds pyrefly and introduces changes in order
to get a clean run via `make check`.

It does seem to have found a few bugs in my
early testing; I stumbled upon it as it was an LSP
option for zed.

One thing it did seem to catch was the session reuse bug
already fixed in upstream where we weren't checking
await_request for None.