Model stubs generation in Makefile
Previously, we fully removed and regenerated
stubs for a bunch of stuff.  Now, only do so
based on changes to compute.wit which should
generally be a good proxy.

This also makes the lint checks much lighter
by not requiring the we build all the examples
just to get stubs.